% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateRanges.R
\name{annotateRanges}
\alias{annotateRanges}
\title{Function searching for overlaps with various annotation ranges.}
\usage{
annotateRanges(searchGRanges = NULL, maxDist = 1000, blacklist = NULL,
  ...)
}
\arguments{
\item{searchGRanges}{Ranges to be searched for overlaps with annotation ranges.}

\item{maxDist}{Maximal distance of a search range from the annotation range to be considered as hit.}

\item{blacklist}{List of regions to be excluded from the analysis (format: chromosome start end).}

\item{...}{Files containing various annotation ranges (format: chromosome start end).}
}
\description{
Function searching for overlaps with various annotation ranges.
}
\author{
David Porubsky
}

