% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLinks.R
\name{collapse.str}
\alias{collapse.str}
\title{This function concatenates strings stored in the list}
\usage{
collapse.str(x, sep = "")
}
\arguments{
\item{x}{List of character vectors}

\item{sep}{Separator to concatenate character vectors}
}
\description{
This function concatenates strings stored in the list
}
\author{
David Porubsky
}

