% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossLink.R
\name{crossLinkR}
\alias{crossLinkR}
\title{Wrapper function for CrossLinkR package}
\usage{
crossLinkR(bamFile, vcfFile, outputfolder = "./CrossLinkR_analysis",
  yieldSize = 1e+05, readLen = 100, min.mapq = 10, min.baseq = 20,
  nCPU = 3, blacklist = NULL)
}
\arguments{
\item{bamFile}{Bamfile with paired-end reads sorted by readName.}

\item{vcfFile}{VCF file containing phased alleles}

\item{outputfolder}{Output directory. If non-existent it will be created.}

\item{yieldSize}{Number of reads to be loaded as a chunk to save memory (Default: 1000000)}

\item{readLen}{Length of the read in bp.}

\item{min.mapq}{Minimum mapping quality of any given read.}

\item{min.baseq}{Minimum base quality to consider SNV for crosslink.}

\item{nCPU}{The number of CPUs to use. Should not be more than available on your machine.}

\item{blacklist}{List of regions to be excluded from the analysis (format: chromosome start end)}
}
\description{
Wrapper function for CrossLinkR package
}
\author{
David Porubsky
}

