% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLinks.R
\name{getLinks}
\alias{getLinks}
\title{Search for links in Hi-C read pairs}
\usage{
getLinks(reads, hap.gr, outputfolder, tmpDIR, readLen = 100, min.mapq = 10,
  min.baseq = 20, ID = "")
}
\arguments{
\item{reads}{An \code{\link{GRanges}} object containing read-pairs sorted by name.}

\item{hap.gr}{An \code{\link{GRanges}} object containing phased SNVs used as a reference}

\item{readLen}{Length of the read in bp.}

\item{min.mapq}{Minimum mapping quality of any given read.}

\item{min.baseq}{Minimum base quality to consider SNV for crosslink.}

\item{ID}{Unique identifier for the analysis. (Default: Bamfile name)}
}
\description{
Search for links in Hi-C read pairs
}
\author{
David Porubsky
}

