% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotCrosslinks.R
\name{hotLinks}
\alias{hotLinks}
\title{Search for hotspots of crosslinked reads}
\usage{
hotLinks(crosslinks = NULL, maxDist = 1000, minCov = 4, pval = NULL,
  ntrials = 100, chrom.lengths)
}
\arguments{
\item{crosslinks}{File containing analyzed links between read-pairs}

\item{maxDist}{Maximal distance to connect neighbouring reads into hotspots.}

\item{minCov}{Minimal coverage to filter hotspot locations.}

\item{pval}{P value cut-off to filter hotspot locations.}

\item{chrom.lengths}{Vector containing lenghts of all analyzed chromosomes (Info in Bam header).}
}
\description{
Search for hotspots of crosslinked reads
}
\author{
David Porubsky
}

