% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotlinkSizeDist}
\alias{plotlinkSizeDist}
\title{Plot size distribution between links}
\usage{
plotlinkSizeDist(linkFile)
}
\arguments{
\item{linkFile}{Filename containing link coordinates}
}
\description{
Plot size distribution between links
}
\author{
David Porubsky
}

