% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotPairwiseTransLinks}
\alias{plotPairwiseTransLinks}
\title{Plot inter-chromosomal crosslinks per chromosomal paur}
\usage{
plotPairwiseTransLinks(linkFile = NULL, chromosome.pair = NULL)
}
\arguments{
\item{linkFile}{Filename containing link coordinates}

\item{chromosome.pair}{Vector containing identifiers of two chromosomes to plot Example: c("chr1", "chr2)}
}
\description{
Plot inter-chromosomal crosslinks per chromosomal paur
}
\author{
David Porubsky
}

