% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotTransLinkCircos}
\alias{plotTransLinkCircos}
\title{Plot inter-chromosomal crosslinks per chromosome}
\usage{
plotTransLinkCircos(linkFile = NULL, genome = "hg38")
}
\arguments{
\item{linkFile}{Filename containing link coordinates}

\item{genome}{Reference genome assembly to use to construct ideogram}
}
\description{
Plot inter-chromosomal crosslinks per chromosome
}
\author{
David Porubsky
}

