% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftRanges.R
\name{shiftRanges}
\alias{shiftRanges}
\title{Shift ranges in certain distance}
\usage{
shiftRanges(ranges = NULL, shiftrange = 1e+07)
}
\arguments{
\item{ranges}{An \code{\link{GRanges}} object to be shifted.}

\item{shiftrange}{The distance, given \code{\link{GRanges}} object will be shifted.}
}
\description{
Shift ranges in certain distance
}
\author{
David Porubsky
}

