% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2Ranges.R
\name{vcf2rangesWH}
\alias{vcf2rangesWH}
\title{Read VCF file into a \code{\link{GRanges}} object}
\usage{
vcf2rangesWH(vcfFile = NULL, genotypeField = 1, filterUnphased = T,
  minimalBlock = 10000)
}
\arguments{
\item{vcfFile}{VCF file containing phased alleles}

\item{genotypeField}{In case of multiple samples phased in a single file (Default: 1)}

\item{filterUnphased}{Filter out unphased alleles (Default: TRUE)}

\item{minimalBlock}{In case of multiple phased blocks, size of the smallest block (Default: 10000)}
}
\description{
Read VCF file into a \code{\link{GRanges}} object
}
\author{
David Porubsky
}

